+++ test/extra/persist/test_champ.cpp
@@ -775,1656 +775,6 @@ TEST_CASE("Test modifying nodes with collisions")
     }
 }
 
-TEST_CASE("Test champ pool conversion, map")
-{
-    using test::new_type;
-    using test::old_type;
-
-    using old_map_t =
-        immer::map<std::string, old_type, immer::persist::xx_hash<std::string>>;
-    using new_map_t =
-        immer::map<std::string, new_type, immer::persist::xx_hash<std::string>>;
-
-    const auto map1 = [] {
-        auto map = old_map_t{};
-        for (auto i = 0; i < 30; ++i) {
-            map =
-                std::move(map).set(fmt::format("x{}x", i), old_type{.data = i});
-        }
-        return map;
-    }();
-    const auto map2 = map1.set("345", old_type{.data = 345});
-
-    auto [pool, map1_id]    = immer::persist::champ::add_to_pool(map1, {});
-    auto map2_id            = immer::persist::node_id{};
-    std::tie(pool, map2_id) = add_to_pool(map2, pool);
-
-    // Confirm that map1 and map2 have structural sharing in the beginning.
-    // For example, "x21x" is stored only once, it's shared.
-    const auto expected_ar = json_t::parse(R"(
-{
-  "value0": [
-    {
-      "children": [
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9
-      ],
-      "collisions": false,
-      "datamap": 2013603464,
-      "nodemap": 2188935188,
-      "values": [
-        {
-          "first": "x13x",
-          "second": {
-            "data": 13,
-            "id": ""
-          }
-        },
-        {
-          "first": "x4x",
-          "second": {
-            "data": 4,
-            "id": ""
-          }
-        },
-        {
-          "first": "x22x",
-          "second": {
-            "data": 22,
-            "id": ""
-          }
-        },
-        {
-          "first": "x28x",
-          "second": {
-            "data": 28,
-            "id": ""
-          }
-        },
-        {
-          "first": "x10x",
-          "second": {
-            "data": 10,
-            "id": ""
-          }
-        },
-        {
-          "first": "x12x",
-          "second": {
-            "data": 12,
-            "id": ""
-          }
-        },
-        {
-          "first": "x9x",
-          "second": {
-            "data": 9,
-            "id": ""
-          }
-        },
-        {
-          "first": "x29x",
-          "second": {
-            "data": 29,
-            "id": ""
-          }
-        },
-        {
-          "first": "x6x",
-          "second": {
-            "data": 6,
-            "id": ""
-          }
-        },
-        {
-          "first": "x17x",
-          "second": {
-            "data": 17,
-            "id": ""
-          }
-        },
-        {
-          "first": "x11x",
-          "second": {
-            "data": 11,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 67125248,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x21x",
-          "second": {
-            "data": 21,
-            "id": ""
-          }
-        },
-        {
-          "first": "x5x",
-          "second": {
-            "data": 5,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 32770,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x25x",
-          "second": {
-            "data": 25,
-            "id": ""
-          }
-        },
-        {
-          "first": "x26x",
-          "second": {
-            "data": 26,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 65539,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x8x",
-          "second": {
-            "data": 8,
-            "id": ""
-          }
-        },
-        {
-          "first": "x16x",
-          "second": {
-            "data": 16,
-            "id": ""
-          }
-        },
-        {
-          "first": "x3x",
-          "second": {
-            "data": 3,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 139264,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x14x",
-          "second": {
-            "data": 14,
-            "id": ""
-          }
-        },
-        {
-          "first": "x18x",
-          "second": {
-            "data": 18,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 1073742080,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x23x",
-          "second": {
-            "data": 23,
-            "id": ""
-          }
-        },
-        {
-          "first": "x0x",
-          "second": {
-            "data": 0,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 2621440,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x15x",
-          "second": {
-            "data": 15,
-            "id": ""
-          }
-        },
-        {
-          "first": "x24x",
-          "second": {
-            "data": 24,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 8224,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x27x",
-          "second": {
-            "data": 27,
-            "id": ""
-          }
-        },
-        {
-          "first": "x1x",
-          "second": {
-            "data": 1,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 8421376,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x7x",
-          "second": {
-            "data": 7,
-            "id": ""
-          }
-        },
-        {
-          "first": "x20x",
-          "second": {
-            "data": 20,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 134234112,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x19x",
-          "second": {
-            "data": 19,
-            "id": ""
-          }
-        },
-        {
-          "first": "x2x",
-          "second": {
-            "data": 2,
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9
-      ],
-      "collisions": false,
-      "datamap": 2013619848,
-      "nodemap": 2188935188,
-      "values": [
-        {
-          "first": "x13x",
-          "second": {
-            "data": 13,
-            "id": ""
-          }
-        },
-        {
-          "first": "x4x",
-          "second": {
-            "data": 4,
-            "id": ""
-          }
-        },
-        {
-          "first": "x22x",
-          "second": {
-            "data": 22,
-            "id": ""
-          }
-        },
-        {
-          "first": "x28x",
-          "second": {
-            "data": 28,
-            "id": ""
-          }
-        },
-        {
-          "first": "x10x",
-          "second": {
-            "data": 10,
-            "id": ""
-          }
-        },
-        {
-          "first": "x12x",
-          "second": {
-            "data": 12,
-            "id": ""
-          }
-        },
-        {
-          "first": "x9x",
-          "second": {
-            "data": 9,
-            "id": ""
-          }
-        },
-        {
-          "first": "x29x",
-          "second": {
-            "data": 29,
-            "id": ""
-          }
-        },
-        {
-          "first": "x6x",
-          "second": {
-            "data": 6,
-            "id": ""
-          }
-        },
-        {
-          "first": "345",
-          "second": {
-            "data": 345,
-            "id": ""
-          }
-        },
-        {
-          "first": "x17x",
-          "second": {
-            "data": 17,
-            "id": ""
-          }
-        },
-        {
-          "first": "x11x",
-          "second": {
-            "data": 11,
-            "id": ""
-          }
-        }
-      ]
-    }
-  ]
-}
-    )");
-
-    REQUIRE(json_t::parse(to_json(pool)) == expected_ar);
-
-    const auto transform_map = [&](const auto& map) {
-        auto result = new_map_t{};
-        for (const auto& [key, value] : map) {
-            result = std::move(result).set(key, convert_old_type(value));
-        }
-        return result;
-    };
-
-    const auto in_pool          = to_input_pool(pool);
-    const auto in_pool_new_type = transform_pool(in_pool, convert_old_type_map);
-    auto loader = immer::persist::champ::container_loader{in_pool_new_type};
-
-    const auto loaded_1 = loader.load(map1_id);
-    const auto loaded_2 = loader.load(map2_id);
-    REQUIRE(loaded_1 == transform_map(map1));
-    REQUIRE(loaded_2 == transform_map(map2));
-
-    SECTION("Loaded maps still share the structure")
-    {
-        auto [pool, id]    = immer::persist::champ::add_to_pool(loaded_1, {});
-        std::tie(pool, id) = add_to_pool(loaded_1, pool);
-        std::tie(pool, id) = add_to_pool(loaded_2, pool);
-
-        // "_21_" is stored only once
-        const auto expected_ar = json_t::parse(R"(
-{
-  "value0": [
-    {
-      "children": [
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9
-      ],
-      "collisions": false,
-      "datamap": 2013603464,
-      "nodemap": 2188935188,
-      "values": [
-        {
-          "first": "x13x",
-          "second": {
-            "data": 13,
-            "data2": "_13_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x4x",
-          "second": {
-            "data": 4,
-            "data2": "_4_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x22x",
-          "second": {
-            "data": 22,
-            "data2": "_22_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x28x",
-          "second": {
-            "data": 28,
-            "data2": "_28_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x10x",
-          "second": {
-            "data": 10,
-            "data2": "_10_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x12x",
-          "second": {
-            "data": 12,
-            "data2": "_12_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x9x",
-          "second": {
-            "data": 9,
-            "data2": "_9_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x29x",
-          "second": {
-            "data": 29,
-            "data2": "_29_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x6x",
-          "second": {
-            "data": 6,
-            "data2": "_6_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x17x",
-          "second": {
-            "data": 17,
-            "data2": "_17_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x11x",
-          "second": {
-            "data": 11,
-            "data2": "_11_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 67125248,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x21x",
-          "second": {
-            "data": 21,
-            "data2": "_21_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x5x",
-          "second": {
-            "data": 5,
-            "data2": "_5_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 32770,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x25x",
-          "second": {
-            "data": 25,
-            "data2": "_25_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x26x",
-          "second": {
-            "data": 26,
-            "data2": "_26_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 65539,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x8x",
-          "second": {
-            "data": 8,
-            "data2": "_8_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x16x",
-          "second": {
-            "data": 16,
-            "data2": "_16_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x3x",
-          "second": {
-            "data": 3,
-            "data2": "_3_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 139264,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x14x",
-          "second": {
-            "data": 14,
-            "data2": "_14_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x18x",
-          "second": {
-            "data": 18,
-            "data2": "_18_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 1073742080,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x23x",
-          "second": {
-            "data": 23,
-            "data2": "_23_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x0x",
-          "second": {
-            "data": 0,
-            "data2": "_0_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 2621440,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x15x",
-          "second": {
-            "data": 15,
-            "data2": "_15_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x24x",
-          "second": {
-            "data": 24,
-            "data2": "_24_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 8224,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x27x",
-          "second": {
-            "data": 27,
-            "data2": "_27_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x1x",
-          "second": {
-            "data": 1,
-            "data2": "_1_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 8421376,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x7x",
-          "second": {
-            "data": 7,
-            "data2": "_7_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x20x",
-          "second": {
-            "data": 20,
-            "data2": "_20_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 134234112,
-      "nodemap": 0,
-      "values": [
-        {
-          "first": "x19x",
-          "second": {
-            "data": 19,
-            "data2": "_19_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x2x",
-          "second": {
-            "data": 2,
-            "data2": "_2_",
-            "id": ""
-          }
-        }
-      ]
-    },
-    {
-      "children": [
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9
-      ],
-      "collisions": false,
-      "datamap": 2013619848,
-      "nodemap": 2188935188,
-      "values": [
-        {
-          "first": "x13x",
-          "second": {
-            "data": 13,
-            "data2": "_13_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x4x",
-          "second": {
-            "data": 4,
-            "data2": "_4_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x22x",
-          "second": {
-            "data": 22,
-            "data2": "_22_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x28x",
-          "second": {
-            "data": 28,
-            "data2": "_28_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x10x",
-          "second": {
-            "data": 10,
-            "data2": "_10_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x12x",
-          "second": {
-            "data": 12,
-            "data2": "_12_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x9x",
-          "second": {
-            "data": 9,
-            "data2": "_9_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x29x",
-          "second": {
-            "data": 29,
-            "data2": "_29_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x6x",
-          "second": {
-            "data": 6,
-            "data2": "_6_",
-            "id": ""
-          }
-        },
-        {
-          "first": "345",
-          "second": {
-            "data": 345,
-            "data2": "_345_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x17x",
-          "second": {
-            "data": 17,
-            "data2": "_17_",
-            "id": ""
-          }
-        },
-        {
-          "first": "x11x",
-          "second": {
-            "data": 11,
-            "data2": "_11_",
-            "id": ""
-          }
-        }
-      ]
-    }
-  ]
-}
-        )");
-        REQUIRE(json_t::parse(to_json(pool)) == expected_ar);
-    }
-}
-
-TEST_CASE("Test champ pool conversion, table")
-{
-    using test::new_type;
-    using test::old_type;
-
-    using old_table_t = immer::table<old_type,
-                                     immer::table_key_fn,
-                                     immer::persist::xx_hash<std::string>>;
-    using new_table_t = immer::table<new_type,
-                                     immer::table_key_fn,
-                                     immer::persist::xx_hash<std::string>>;
-
-    const auto table1 = [] {
-        auto table = old_table_t{};
-        for (auto i = 0; i < 30; ++i) {
-            table = std::move(table).insert(old_type{
-                .id   = fmt::format("q{}q", i),
-                .data = i,
-            });
-        }
-        return table;
-    }();
-    const auto table2 = table1.insert(old_type{
-        .id   = "345",
-        .data = 345,
-    });
-
-    auto [pool, table1_id]    = immer::persist::champ::add_to_pool(table1, {});
-    auto table2_id            = immer::persist::node_id{};
-    std::tie(pool, table2_id) = add_to_pool(table2, pool);
-
-    // Confirm that table1 and table2 have structural sharing in the beginning.
-    // "q2q" is stored only once.
-    const auto expected_ar = json_t::parse(R"(
-  {
-    "value0": [
-      {
-        "children": [
-          1,
-          2,
-          3,
-          4,
-          5,
-          6,
-          7,
-          8,
-          9,
-          10
-        ],
-        "collisions": false,
-        "datamap": 3225456713,
-        "nodemap": 889282612,
-        "values": [
-          {
-            "data": 25,
-            "id": "q25q"
-          },
-          {
-            "data": 26,
-            "id": "q26q"
-          },
-          {
-            "data": 16,
-            "id": "q16q"
-          },
-          {
-            "data": 3,
-            "id": "q3q"
-          },
-          {
-            "data": 12,
-            "id": "q12q"
-          },
-          {
-            "data": 7,
-            "id": "q7q"
-          },
-          {
-            "data": 10,
-            "id": "q10q"
-          },
-          {
-            "data": 24,
-            "id": "q24q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 539492352,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 2,
-            "id": "q2q"
-          },
-          {
-            "data": 19,
-            "id": "q19q"
-          },
-          {
-            "data": 15,
-            "id": "q15q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 4325376,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 9,
-            "id": "q9q"
-          },
-          {
-            "data": 13,
-            "id": "q13q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 262272,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 6,
-            "id": "q6q"
-          },
-          {
-            "data": 27,
-            "id": "q27q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 136,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 28,
-            "id": "q28q"
-          },
-          {
-            "data": 29,
-            "id": "q29q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 8194,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 1,
-            "id": "q1q"
-          },
-          {
-            "data": 4,
-            "id": "q4q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 2147745792,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 20,
-            "id": "q20q"
-          },
-          {
-            "data": 17,
-            "id": "q17q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 537395712,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 0,
-            "id": "q0q"
-          },
-          {
-            "data": 14,
-            "id": "q14q"
-          },
-          {
-            "data": 21,
-            "id": "q21q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 536870920,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 8,
-            "id": "q8q"
-          },
-          {
-            "data": 11,
-            "id": "q11q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 32800,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 5,
-            "id": "q5q"
-          },
-          {
-            "data": 18,
-            "id": "q18q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 12582912,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 22,
-            "id": "q22q"
-          },
-          {
-            "data": 23,
-            "id": "q23q"
-          }
-        ]
-      },
-      {
-        "children": [
-          1,
-          2,
-          3,
-          4,
-          5,
-          6,
-          12,
-          8,
-          9,
-          10
-        ],
-        "collisions": false,
-        "datamap": 3225456713,
-        "nodemap": 889282612,
-        "values": [
-          {
-            "data": 25,
-            "id": "q25q"
-          },
-          {
-            "data": 26,
-            "id": "q26q"
-          },
-          {
-            "data": 16,
-            "id": "q16q"
-          },
-          {
-            "data": 3,
-            "id": "q3q"
-          },
-          {
-            "data": 12,
-            "id": "q12q"
-          },
-          {
-            "data": 7,
-            "id": "q7q"
-          },
-          {
-            "data": 10,
-            "id": "q10q"
-          },
-          {
-            "data": 24,
-            "id": "q24q"
-          }
-        ]
-      },
-      {
-        "children": [],
-        "collisions": false,
-        "datamap": 570950144,
-        "nodemap": 0,
-        "values": [
-          {
-            "data": 345,
-            "id": "345"
-          },
-          {
-            "data": 0,
-            "id": "q0q"
-          },
-          {
-            "data": 14,
-            "id": "q14q"
-          },
-          {
-            "data": 21,
-            "id": "q21q"
-          }
-        ]
-      }
-    ]
-  }
-    )");
-
-    REQUIRE(json_t::parse(to_json(pool)) == expected_ar);
-
-    const auto transform_table = [&](const auto& table) {
-        auto result = new_table_t{};
-        for (const auto& item : table) {
-            result = std::move(result).insert(convert_old_type(item));
-        }
-        return result;
-    };
-
-    const auto in_pool = to_input_pool(pool);
-
-    SECTION("Invalid conversion, ID is corrupted")
-    {
-        const auto badly_convert_old_type = hana::overload(
-            [](const old_type& val) {
-                return new_type{
-                    .id    = val.id + "OOPS",
-                    .data  = val.data,
-                    .data2 = fmt::format("_{}_", val.data),
-                };
-            },
-            [](immer::persist::target_container_type_request) {
-                return immer::table<new_type>{};
-            });
-        const auto in_pool_new_type =
-            transform_pool(in_pool, badly_convert_old_type);
-        auto loader = immer::persist::champ::container_loader{in_pool_new_type};
-        REQUIRE_THROWS_AS(
-            loader.load(table1_id),
-            immer::persist::champ::hash_validation_failed_exception);
-        REQUIRE_THROWS_AS(
-            loader.load(table2_id),
-            immer::persist::champ::hash_validation_failed_exception);
-    }
-
-    SECTION("Valid conversion, ID is not changed")
-    {
-        const auto in_pool_new_type =
-            transform_pool(in_pool, convert_old_type_table);
-        auto loader = immer::persist::champ::container_loader{in_pool_new_type};
-
-        const auto loaded_1 = loader.load(table1_id);
-        const auto loaded_2 = loader.load(table2_id);
-        REQUIRE(loaded_1 == transform_table(table1));
-        REQUIRE(loaded_2 == transform_table(table2));
-
-        SECTION("Loaded tables still share the structure")
-        {
-            auto [pool, id] = immer::persist::champ::add_to_pool(loaded_1, {});
-            std::tie(pool, id) = add_to_pool(loaded_1, pool);
-            std::tie(pool, id) = add_to_pool(loaded_2, pool);
-
-            // For example, "q17q" is stored only once
-            const auto expected_ar = json_t::parse(R"(
-{
-  "value0": [
-    {
-      "children": [
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        7,
-        8,
-        9,
-        10
-      ],
-      "collisions": false,
-      "datamap": 3225456713,
-      "nodemap": 889282612,
-      "values": [
-        {
-          "data": 25,
-          "data2": "_25_",
-          "id": "q25q"
-        },
-        {
-          "data": 26,
-          "data2": "_26_",
-          "id": "q26q"
-        },
-        {
-          "data": 16,
-          "data2": "_16_",
-          "id": "q16q"
-        },
-        {
-          "data": 3,
-          "data2": "_3_",
-          "id": "q3q"
-        },
-        {
-          "data": 12,
-          "data2": "_12_",
-          "id": "q12q"
-        },
-        {
-          "data": 7,
-          "data2": "_7_",
-          "id": "q7q"
-        },
-        {
-          "data": 10,
-          "data2": "_10_",
-          "id": "q10q"
-        },
-        {
-          "data": 24,
-          "data2": "_24_",
-          "id": "q24q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 539492352,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 2,
-          "data2": "_2_",
-          "id": "q2q"
-        },
-        {
-          "data": 19,
-          "data2": "_19_",
-          "id": "q19q"
-        },
-        {
-          "data": 15,
-          "data2": "_15_",
-          "id": "q15q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 4325376,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 9,
-          "data2": "_9_",
-          "id": "q9q"
-        },
-        {
-          "data": 13,
-          "data2": "_13_",
-          "id": "q13q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 262272,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 6,
-          "data2": "_6_",
-          "id": "q6q"
-        },
-        {
-          "data": 27,
-          "data2": "_27_",
-          "id": "q27q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 136,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 28,
-          "data2": "_28_",
-          "id": "q28q"
-        },
-        {
-          "data": 29,
-          "data2": "_29_",
-          "id": "q29q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 8194,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 1,
-          "data2": "_1_",
-          "id": "q1q"
-        },
-        {
-          "data": 4,
-          "data2": "_4_",
-          "id": "q4q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 2147745792,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 20,
-          "data2": "_20_",
-          "id": "q20q"
-        },
-        {
-          "data": 17,
-          "data2": "_17_",
-          "id": "q17q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 537395712,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 0,
-          "data2": "_0_",
-          "id": "q0q"
-        },
-        {
-          "data": 14,
-          "data2": "_14_",
-          "id": "q14q"
-        },
-        {
-          "data": 21,
-          "data2": "_21_",
-          "id": "q21q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 536870920,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 8,
-          "data2": "_8_",
-          "id": "q8q"
-        },
-        {
-          "data": 11,
-          "data2": "_11_",
-          "id": "q11q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 32800,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 5,
-          "data2": "_5_",
-          "id": "q5q"
-        },
-        {
-          "data": 18,
-          "data2": "_18_",
-          "id": "q18q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 12582912,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 22,
-          "data2": "_22_",
-          "id": "q22q"
-        },
-        {
-          "data": 23,
-          "data2": "_23_",
-          "id": "q23q"
-        }
-      ]
-    },
-    {
-      "children": [
-        1,
-        2,
-        3,
-        4,
-        5,
-        6,
-        12,
-        8,
-        9,
-        10
-      ],
-      "collisions": false,
-      "datamap": 3225456713,
-      "nodemap": 889282612,
-      "values": [
-        {
-          "data": 25,
-          "data2": "_25_",
-          "id": "q25q"
-        },
-        {
-          "data": 26,
-          "data2": "_26_",
-          "id": "q26q"
-        },
-        {
-          "data": 16,
-          "data2": "_16_",
-          "id": "q16q"
-        },
-        {
-          "data": 3,
-          "data2": "_3_",
-          "id": "q3q"
-        },
-        {
-          "data": 12,
-          "data2": "_12_",
-          "id": "q12q"
-        },
-        {
-          "data": 7,
-          "data2": "_7_",
-          "id": "q7q"
-        },
-        {
-          "data": 10,
-          "data2": "_10_",
-          "id": "q10q"
-        },
-        {
-          "data": 24,
-          "data2": "_24_",
-          "id": "q24q"
-        }
-      ]
-    },
-    {
-      "children": [],
-      "collisions": false,
-      "datamap": 570950144,
-      "nodemap": 0,
-      "values": [
-        {
-          "data": 345,
-          "data2": "_345_",
-          "id": "345"
-        },
-        {
-          "data": 0,
-          "data2": "_0_",
-          "id": "q0q"
-        },
-        {
-          "data": 14,
-          "data2": "_14_",
-          "id": "q14q"
-        },
-        {
-          "data": 21,
-          "data2": "_21_",
-          "id": "q21q"
-        }
-      ]
-    }
-  ]
-}
-        )");
-            REQUIRE(json_t::parse(to_json(pool)) == expected_ar);
-        }
-    }
-}
-
 namespace {
 struct old_hash
 {
