+++ test/extra/persist/test_special_pool_auto.cpp
@@ -304,72 +304,6 @@ struct new_app_type
 
 } // namespace
 
-TEST_CASE("Test conversion with auto-pool")
-{
-    const auto vec1 = test::vector_one<old_type>{
-        old_type{.data = 123},
-        old_type{.data = 234},
-    };
-    const auto vec2 = vec1.push_back(old_type{.data = 345});
-
-    const auto map1 = [] {
-        auto map = map_t<old_type>{};
-        for (auto i = 0; i < 30; ++i) {
-            map =
-                std::move(map).set(fmt::format("x{}x", i), old_type{.data = i});
-        }
-        return map;
-    }();
-    const auto map2 = map1.set("345", old_type{.data = 345});
-
-    // Prepare a value of the old type that uses some structural sharing
-    // internally.
-    const auto value = old_app_type{
-        .vec  = vec1,
-        .vec2 = vec2,
-        .map  = map1,
-        .map2 = map2,
-        .table =
-            {
-                old_type{"_51_", 51},
-                old_type{"_52_", 52},
-                old_type{"_53_", 53},
-            },
-    };
-
-    const auto old_pool_types =
-        immer::persist::get_pools_for_type(hana::type_c<old_app_type>);
-
-    const auto json_str =
-        immer::persist::to_json_with_auto_pool(value, old_pool_types);
-    // REQUIRE(json_str == "");
-
-    // Describe how to go from the old pool to the desired new pool.
-    // Convert all old pools with convert_old_type.
-    const auto pools_conversions =
-        hana::make_map(hana::make_pair(hana::type_c<test::vector_one<old_type>>,
-                                       test::convert_old_type),
-                       hana::make_pair(hana::type_c<map_t<old_type>>,
-                                       test::convert_old_type_map),
-                       hana::make_pair(hana::type_c<table_t<old_type>>,
-                                       test::convert_old_type_table));
-
-    // Having a JSON from serializing old_app_type and a conversion function,
-    // we need to somehow load new_app_type.
-    const new_app_type full_load =
-        immer::persist::from_json_with_auto_pool_with_conversion<new_app_type,
-                                                                 old_app_type>(
-            json_str, pools_conversions, old_pool_types);
-
-    {
-        REQUIRE(full_load.vec == transform_vec(value.vec));
-        REQUIRE(full_load.vec2 == transform_vec(value.vec2));
-        REQUIRE(full_load.map == transform_map(value.map));
-        REQUIRE(full_load.map2 == transform_map(value.map2));
-        REQUIRE(full_load.table == transform_table(value.table));
-    }
-}
-
 namespace champ_test {
 
 struct value_two;
