+++ test/extra/persist/test_vectors.cpp
@@ -1545,167 +1545,6 @@ TEST_CASE("Can't load saved flex vector with relaxed nodes as strict")
     }
 }
 
-TEST_CASE("Test vector pool conversion")
-{
-    using test::new_type;
-    using test::old_type;
-
-    const auto vec1 = test::vector_one<old_type>{
-        old_type{.data = 123},
-        old_type{.data = 234},
-    };
-    const auto vec2 = vec1.push_back(old_type{.data = 345});
-
-    auto pool               = immer::persist::rbts::make_output_pool_for(vec1);
-    auto vec1_id            = immer::persist::container_id{};
-    std::tie(pool, vec1_id) = add_to_pool(vec1, pool);
-    auto vec2_id            = immer::persist::container_id{};
-    std::tie(pool, vec2_id) = add_to_pool(vec2, pool);
-
-    // Confirm that vec1 and vec2 have structural sharing in the beginning.
-    const auto expected_pool = json_t::parse(R"(
-{
-  "value0": {
-    "inners": [
-      {
-        "key": 0,
-        "value": {
-          "children": [],
-          "relaxed": false
-        }
-      },
-      {
-        "key": 2,
-        "value": {
-          "children": [
-            1
-          ],
-          "relaxed": false
-        }
-      }
-    ],
-    "leaves": [
-      {
-        "key": 1,
-        "value": [
-          {
-            "data": 123,
-            "id": ""
-          },
-          {
-            "data": 234,
-            "id": ""
-          }
-        ]
-      },
-      {
-        "key": 3,
-        "value": [
-          {
-            "data": 345,
-            "id": ""
-          }
-        ]
-      }
-    ],
-    "vectors": [
-      {
-        "root": 0,
-        "tail": 1
-      },
-      {
-        "root": 2,
-        "tail": 3
-      }
-    ]
-  }
-}
-    )");
-
-    REQUIRE(json_t::parse(to_json(pool)) == expected_pool);
-
-    const auto in_pool          = to_input_pool(pool);
-    const auto in_pool_new_type = transform_pool(in_pool, convert_old_type);
-    auto loader =
-        make_loader_for(test::vector_one<new_type>{}, in_pool_new_type);
-
-    const auto loaded_1 = loader.load(vec1_id);
-    const auto loaded_2 = loader.load(vec2_id);
-    REQUIRE(loaded_1 == transform_vec(vec1));
-    REQUIRE(loaded_2 == transform_vec(vec2));
-
-    SECTION("Loaded vectors still share the structure")
-    {
-        auto pool = immer::persist::rbts::make_output_pool_for(loaded_1);
-        auto id   = immer::persist::container_id{};
-        std::tie(pool, id) = add_to_pool(loaded_1, pool);
-        std::tie(pool, id) = add_to_pool(loaded_2, pool);
-
-        const auto expected_pool = json_t::parse(R"(
-{
-  "value0": {
-    "inners": [
-      {
-        "key": 0,
-        "value": {
-          "children": [],
-          "relaxed": false
-        }
-      },
-      {
-        "key": 2,
-        "value": {
-          "children": [
-            1
-          ],
-          "relaxed": false
-        }
-      }
-    ],
-    "leaves": [
-      {
-        "key": 1,
-        "value": [
-          {
-            "data": 123,
-            "data2": "_123_",
-            "id": ""
-          },
-          {
-            "data": 234,
-            "data2": "_234_",
-            "id": ""
-          }
-        ]
-      },
-      {
-        "key": 3,
-        "value": [
-          {
-            "data": 345,
-            "data2": "_345_",
-            "id": ""
-          }
-        ]
-      }
-    ],
-    "vectors": [
-      {
-        "root": 0,
-        "tail": 1
-      },
-      {
-        "root": 2,
-        "tail": 3
-      }
-    ]
-  }
-}
-        )");
-        REQUIRE(json_t::parse(to_json(pool)) == expected_pool);
-    }
-}
-
 TEST_CASE("Vector: converting loader can handle exceptions")
 {
     const auto vec                = example_vector{1, 2, 3};
